backend/mongoapi/requirements.txt
fastapi=0.87.0=pyhd8ed1ab_0
motor=3.1.1=pyhd8ed1ab_0
uvicorn=0.20.0=py311h38be061_1
python-dotenv=0.21.0=pyhd8ed1ab_0

